changes in documenatiton, style, etc.
