+++ packages/extester-runner/src/ui-test/resources/example-project/.vscodeignore
@@ -1,0 +1,1 @@
+temporary-folder-test-04/
