+++ packages/extester-runner/src/ui-test/utils/testUtils.ts
@@ -1,14 +1,16 @@
 import path from 'path';
-import { ActivityBar, By, ExtensionsViewItem, ExtensionsViewSection, WebDriver } from 'vscode-extension-tester';
+import { ActivityBar, BottomBarPanel, By, EditorView, ExtensionsViewItem, ExtensionsViewSection, TerminalView, ViewSection, WebDriver, Workbench } from 'vscode-extension-tester';
 
 export const RESOURCES: string = path.resolve('src', 'ui-test', 'resources');
 export const EXAMPLE_PROJECT: string = path.join(RESOURCES, 'example-project');
 
 // actions
+export const ACTIONS_FOLDER = process.platform === 'win32' ? 'src\\actions' : 'src/actions';
 export const CREATE_LOG_FILE = 'createLog.test.ts';
 export const CREATE_SCREENSHOT_FILE = 'createScreenshot.test.ts';
 
 // parser
+export const PARSER_FOLDER = process.platform === 'win32' ? 'src\\parser' : 'src/parser';
 export const COMPLEX_FILE = 'complexFile.test.ts';
 export const MULTIPLE_ROOT_DESCRIBES = 'multipleRootDescribes.test.ts';
 export const SIMPLE_FILE = 'simpleFile.test.ts';
@@ -33,6 +35,9 @@ export const ADDITIONAL_ARGS_SETTINGS_ID = 'extesterRunner.additionalArgs';
 // logs
 export const HIDE_EMPTY_LOGS_FOLDERS_SETTINGS_ID = 'extesterRunner.hideEmptyLogFolders';
 
+// defaults
+export const TEST_FILE_GLOB_SETTINGS_DEFAULT = '**/ui-test/**/*.test.ts';
+
 export const TESTS_VIEW = 'UI Tests';
 export const TESTS_VIEW_NO_TESTS = 'No tests';
 export const TEST_VIEW_REFRESH_BTN = 'Refresh Tests';
@@ -48,12 +53,11 @@ export const LOGS_VIEW_REFRESH_BTN = 'Refresh Logs';
 export const COLLAPSE_ALL_BTN = 'Collapse All';
 
 /**
- * Wait until required extension is activated.
- *
- * @param driver WebDriver.
- * @param displayName Name of extension.
- * @param timeout Timeout for dynamic wait.
- * @param interval Delay between individual checks.
+ * Waits until a specific extension is activated in VS Code
+ * @param driver - WebDriver instance
+ * @param displayName - Name of the extension to wait for
+ * @param timeout - Maximum time to wait in milliseconds (default: 150000)
+ * @param interval - Time between checks in milliseconds (default: 500)
  */
 export async function waitUntilExtensionIsActivated(driver: WebDriver, displayName: string, timeout = 150000, interval = 500): Promise<void> {
 	await driver.wait(
@@ -67,10 +71,9 @@ export async function waitUntilExtensionIsActivated(driver: WebDriver, displayNa
 }
 
 /**
- * Checks, if extension is activated.
- *
- * @param displayName Name of extension.
- * @returns true/false
+ * Checks if a specific extension is activated in VS Code
+ * @param displayName - Name of the extension to check
+ * @returns Promise resolving to true if extension is activated, false otherwise
  */
 export async function extensionIsActivated(displayName: string): Promise<boolean> {
 	try {
@@ -88,4 +91,74 @@ export async function extensionIsActivated(displayName: string): Promise<boolean
 		console.error('Error checking extension activation status:', err);
 		return false;
 	}
+}
+
+/**
+ * Retrieves a specific section from the ExTester Runner view
+ * @param sectionIndex - Index of the section to retrieve
+ * @returns Promise resolving to the requested ViewSection
+ * @throws Error if section is not found or index is out of bounds
+ */
+export async function getSection(sectionIndex: number): Promise<ViewSection> {
+	const runnerView = await (await new ActivityBar().getViewControl(EXTESTER_RUNNER))?.openView();
+	const content = await (await runnerView?.getContent())?.getSections();
+	if (!content) {
+		throw new Error('Content sections not found');
+	}
+	if (content.length <= sectionIndex) {
+		throw new Error(`Content section at index ${sectionIndex} is unavailable`);
+	}
+	return content[sectionIndex];
+}
+
+/**
+ * Updates a VS Code setting with a new value
+ * @param settingId - ID of the setting to update
+ * @param value - New value to set
+ * @throws Error if setting is not found
+ */
+export async function updateSettings(settingId: string, value: string): Promise<void> {
+	const workbench = new Workbench();
+	const settingsEditor = await workbench.openSettings();
+	const settingControl = await settingsEditor.findSettingByID(settingId);
+	if (!settingControl) {
+		throw new Error(`Setting not found: ${TEST_FILE_GLOB_SETTINGS_ID}`);
+	}
+	await settingControl.setValue(value);
+	const editorView = new EditorView();
+	await editorView.closeAllEditors();
+}
+
+/**
+ * Waits until specific text appears in the terminal
+ * @param driver - WebDriver instance
+ * @param text - Text to wait for
+ * @param timeout - Maximum time to wait in milliseconds (default: 120000)
+ * @param interval - Time between checks in milliseconds (default: 2000)
+ */
+export async function waitUntilTerminalHasText(driver: WebDriver, text: string, timeout = 120000, interval = 2000): Promise<void> {
+	await driver.sleep(interval);
+	await driver.wait(
+		async function () {
+			try {
+				const terminal = await activateTerminalView();
+				const terminalText = await terminal.getText();
+				return terminalText.includes(text);
+			} catch (err) {
+				return false;
+			}
+		},
+		timeout,
+		`Unable to find in terminal text: ${text}`,
+		interval,
+	);
+}
+
+/**
+ * Activates the terminal view in VS Code
+ * @returns Promise resolving to the TerminalView instance
+ */
+async function activateTerminalView(): Promise<TerminalView> {
+	await new Workbench().executeCommand('Terminal: Focus on Terminal View');
+	return await new BottomBarPanel().openTerminalView();
 }
